Extract CommandLineArgumentProviders to top-level classes
